feat: main-service: - пакет request: создан дополнительные методы для получения запросов
- пакет event: создан слой репозитория для работы с БД, частично реализована сервисная логика

+++ main-service/src/main/java/ru/yandex/practicum/mainserver/event/dto/AdminUpdateEventRequest.java
@@ -23,7 +23,7 @@ public class AdminUpdateEventRequest {
     private LocalDateTime eventDate;
     private Location location;
     private Boolean paid;
-    private Long participantLimit;
+    private Integer participantLimit;
     private Boolean requestModeration;
     private LocalDateTime createdOn;
